Update ArduinoPID.h
reverted type of error and lastError to int32 because multiplication with another int16 and storage in a int32 output result does not automatically promote them to 32bit prior to the multiplication

+++ ArduinoPID.h
@@ -77,8 +77,8 @@ class ArduinoPID{
 	
 	volatile int32_t setpoint;
 	volatile int32_t measurement;
-	int16_t error;
-	int16_t lastError;			//used if FILTRED_DERIVATIVE == false
+	int32_t error;
+	int32_t lastError;			//used if FILTRED_DERIVATIVE == false
 	int32_t lastMeasurement;	//used if FILTRED_DERIVATIVE == false
 	
 	DerivativeFiltered derFiltered;
